LegalForce class created, first method is_weekend() added
The LegalForce class is supposed to be the cornerstone of the whole project.

Its main intended purpose in the future is to be able to receive a date of proper delivery of the court decision to the party to the lawsuit. Based on that, it should count the day of court decision becoming legally bounding. This will have to deal with weekends and public holidays in the Czech Republic (including changing Easters date).

At this time, only is_weekend() method was created.