LegalForce -> is_workday() added; stores holidays in the instance list; switched logic of raising TypeErrors
Because later on we might need to check for public_holiday in iterative manner, we want to store list of holidays in the instance attribute instead of obtaining it every time with method that creates the list from scratch.

czech_public_holidays() methods was kept, but is now used to set instance holiday attribute in the __init__()

New year holiday of the next year was added to the list of holidays to cover legal force computation at the end of the current year. This solution would not be smart, should the 15-day deadline for appeal ever change or new public holidays were added. In such case, more "dynamic" approach of calculating holidays of the next year should be considered. This is highly unlikely though, so this simple solution will do.

The logic of raising value errors if given_date was not valid was switched - now raising the error comes first and comes return next. This should make the code more readable.

is_workday() method added. Simply, date is a workday when it is not weekend or public holiday.