adding shift_to_workday() method to LegalForce, adding boolean flags for (non) skipping the date validation and holiday recalculation
shift_to_workday() - this new method takes the given date and makes sure it falls on workday. If it does not, it shifts given date one day forward until ordinary workday is found.

Since consecutive holidays do occur in the Czech republic, sometimes several iterations of shifting is needed (for example if given date is 2025-12-24, next workday is 2025-12-29 - five days ahead).

In such cases, recalculating holidays and validating date in each iteration are not necessary. For this purpose, hidden attributes were added to serve as a flags to turn off validation and recalculation. After shifting is over, flags are turned back on.