LegalForce class changed radically - most of the methods changed to static
Throughout the development an important question arose. What advantage does it bring to bind the date related operations to the instance of a class?

None, it showed, quite the opposite. The current approach makes methods more rigid and less usable for general purpose. Most of the data related methods can be easily treated as static methods.

The LegalForce class was kept, but merely as an organizer for related methods. No instance will be needed in most cases.