creating LegalForceManager for interacting with the user
The LegalForce class serves merely as a calculator for the calendar dates related to the legal force of the court decision.

Because of this, it can be used also by other programs, for example to compute the date of legal force for a large database of delivery dates.

For the purpose of interaction with the user and accepting user inputs, the interactive logic was split into separate class called LegalForceManager.

LegalForceManager now has one static method called interact().

 The interact() method serves as a control of the flow of user-based interaction from the command line. It prompts user for date input, uses LegalForce to calculate the outcome and print the result back into console. It does so repetitively until user decides to quit.

  This is cyclical style of interaction is useful, since it is very common technique for court officers to count many legal force dates at once, after they pile up in his office.

  Separating the calculation logic from the user interaction aims to align the code with the correct OOP principles (naimly SRP).

 adding the main guard for executing the LegalForceManager interaction when running the script directly.